pyproject.toml
[project]
name = "kisskh-downloader"
version = "0.1.0"
description = "Simple downloaded for https://kisskh.co/"
readme = "README.md"
authors = [{name = "Debakar Roy", email = "allinonedibakar@gmail.com"}]
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.3",
    "requests>=2.28.2",
    "pydantic>=1.10.5",
    "validators>=0.20.0",
    "yt-dlp>=2023.2.17",
    "pysrt>=1.1.2",
    "python-dotenv>=1.0.0",
    "cryptography>=41.0.7",
]

[project.optional-dependencies]
dev = [
    "datamodel-code-generator[http]>=0.17.1",
    "flake8>=6.0.0",
    "flake8-bugbear>=23.2.13",
    "flake8-black>=0.3.6",
    "flake8-bandit>=4.1.1",
    "flake8-comprehensions>=3.10.1",
    "flake8-isort>=6.0.0",
    "pep8-naming>=0.13.3",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.1",
    "black>=23.1.0",
    "types-requests>=2.28.11.15",
    "pytest>=9.0.2",
]

[project.scripts]
kisskh = "kisskh_downloader.cli:kisskh"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.black]
line-length = 120
